Add reusable the encoder and decoder test class
